+++ conversejs/loc.keys.js
@@ -59,7 +59,8 @@ const locKeys = [
   'moderator_note_create_for_participant',
   'moderator_note_search_for_participant',
   'moderator_note_filters',
-  'moderator_note_original_nick'
+  'moderator_note_original_nick',
+  'search_occupant_message'
 ]
 
 module.exports = locKeys